nii2mnc: count voxels with size_t in do_data_conversion()
do_data_conversion() looped with "int i < (int) nelem". With more than
2^31 - 1 elements (for example a large 4-D series converted with -short,
about 8 GB of float input), (int) nelem is negative, the loop does not
run, and the uninitialized new buffer is written as the image.

Use size_t for the counter, as find_data_range() does.

Refs #130.

+++ conversion/nifti1/nii2mnc.c
@@ -136,7 +136,7 @@ do_data_conversion(nifti_image *nii_ptr,
   double input_rng = (img_rrange[1] - img_rrange[0]);
   double output_min = img_vrange[0];
   double output_rng = (img_vrange[1] - img_vrange[0]);
-  int i;
+  size_t i;
 
   if (new_data == NULL) {
     fprintf(stderr, "ERROR: Out of memory.\n");
@@ -150,7 +150,7 @@ do_data_conversion(nifti_image *nii_ptr,
     img_vrange[1] = img_rrange[1];
   }
 
-  for (i = 0; i < (int) nelem; i++) {
+  for (i = 0; i < nelem; i++) {
     double value;
 
     switch (img_mtype) {
